// server/@types/manage-recalls-api/models/PartBRecordRequest.ts
/* istanbul ignore file */
/* tslint:disable */
/* eslint-disable */

export type PartBRecordRequest = {
    details: string;
    partBReceivedDate: string;
    partBFileName: string;
    partBFileContent: string;
    emailFileName: string;
    emailFileContent: string;
    oasysFileName?: string;
    oasysFileContent?: string;
};